requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
flask-cors==6.0.0
gunicorn==23.0.0
idna==3.10
img2pdf==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
packaging==25.0
pikepdf==9.7.0
pillow==11.2.1
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
